Fix flow and lint, again
