+++ src/geo/projection/albers.js
@@ -1,5 +1,6 @@
 // @flow
 import LngLat from '../lng_lat.js';
+import makeTileTransform from './tile_transform.js';
 
 const albersConstants = {
     refLng: -96,
@@ -51,24 +52,34 @@ function unproject(x, y, constants) {
     return new LngLat(lng, lat);
 }
 
+function albersProject(lng, lat) {
+    return project(lng, lat, albersConstants);
+}
+
+function alaskaProject(lng, lat) {
+    return project(lng, lat, alaskaConstants);
+}
+
 export const albers = {
     name: 'albers',
     range: [3.5, 7],
-    project: (lng, lat) => {
-        return project(lng, lat, albersConstants);
+    project: (lng: number, lat: number) => {
+        return albersProject(lng, lat);
     },
-    unproject: (x, y) => {
+    unproject: (x: number, y: number) => {
         return unproject(x, y, albersConstants);
-    }
+    },
+    tileTransform: makeTileTransform(albersProject)
 };
 
 export const alaska = {
     name: 'alaska',
     range: [4, 7],
-    project: (lng, lat) => {
-        return project(lng, lat, alaskaConstants);
+    project: (lng: number, lat: number) => {
+        return alaskaProject(lng, lat);
     },
-    unproject: (x, y) => {
+    unproject: (x: number, y: number) => {
         return unproject(x, y, alaskaConstants);
-    }
+    },
+    tileTransform: makeTileTransform(alaskaProject)
 };
