+++ src/geo/projection/mercator.js
@@ -1,13 +1,18 @@
 // @flow
 import MercatorCoordinate, {mercatorXfromLng, mercatorYfromLat} from '../mercator_coordinate.js';
+import makeTileTransform from './tile_transform.js';
+
+function project(lat: number, lng: number) {
+    const x = mercatorXfromLng(lng);
+    const y = mercatorYfromLat(lat);
+
+    return {x, y};
+}
 
 export default {
     name: 'mercator',
-    project: (lng, lat) => {
-        const x = mercatorXfromLng(lng);
-        const y = mercatorYfromLat(lat);
-
-        return {x, y};
-    },
-    unproject: (x, y) => new MercatorCoordinate(x, y).toLngLat()
+    range: [],
+    project,
+    unproject: (x: number, y: number) => new MercatorCoordinate(x, y).toLngLat(),
+    tileTransform: makeTileTransform(project)
 };
