+++ src/geo/projection/winkelTripel.js
@@ -4,5 +4,6 @@ export default {
     name: 'winkel',
     range: [3.5, 7],
     project: () => {},
-    unproject: () => {}
+    unproject: () => {},
+    tileTransform: () => {}
 };
